rocm-ck: relocate FMHA BWD headers to include/rocm_ck/ops/fmha_bwd/
Move spec, api, and common headers out of the example directory into
the shared include tree so they can be compiled and tested by a
host-only g++ toolchain without HIP.

The registry (variant tables + findVariant) stays in the example
directory since it defines example-specific kernel configurations.

Changes during relocation:
- Replace dim3 with GridDim in _api.hpp headers (drops hip_runtime.h)
- Drop aggregate rocm_fmha_bwd_api.hpp; consumers include per-family
  _api.hpp headers directly
- Registry now includes _spec.hpp (not _api.hpp), removing its
  transitive HIP dependency
- Prune static_asserts to one compile canary per _spec.hpp; the full
  validation suite will move to dedicated g++ tests in a later phase
- Update all consumers (_dev.hpp, main.cpp, CMakeLists.txt) to use
  new <rocm_ck/ops/fmha_bwd/...> include paths
