+++ projects/composablekernel/experimental/rocm_ck/examples/06_rocm_ck_fmha_bwd/rocm_fmha_bwd_ograd_dot_o_dev.hpp
@@ -19,10 +19,10 @@
 
 #ifndef __HIP_DEVICE_COMPILE__
 #error "rocm_fmha_bwd_ograd_dot_o_dev.hpp requires device compilation." \
-       " Host code should include rocm_fmha_bwd_ograd_dot_o_api.hpp."
+       " Host code should include <rocm_ck/ops/fmha_bwd/ograd_dot_o_api.hpp>."
 #endif
 
-#include "rocm_fmha_bwd_ograd_dot_o_spec.hpp"
+#include <rocm_ck/ops/fmha_bwd/ograd_dot_o_spec.hpp>
 
 #include <rocm_ck/args.hpp>
 #include <rocm_ck/ck_type_map.hpp>
