+++ projects/composablekernel/experimental/rocm_ck/include/rocm_ck/ops/fmha_bwd/convert_dq_api.hpp
@@ -14,13 +14,13 @@
 #pragma once
 
 #ifdef __HIP_DEVICE_COMPILE__
-#error "rocm_fmha_bwd_convert_dq_api.hpp is host-only." \
+#error "convert_dq_api.hpp is host-only." \
        " Device code should include rocm_fmha_bwd_convert_dq_dev.hpp."
 #endif
 
-#include "rocm_fmha_bwd_convert_dq_spec.hpp"
+#include <rocm_ck/ops/fmha_bwd/convert_dq_spec.hpp>
 
-#include <hip/hip_runtime.h>
+#include <rocm_ck/grid_dim.hpp>
 
 namespace rocm_ck {
 
@@ -30,14 +30,14 @@ namespace rocm_ck {
 
 /// Compute the launch grid for ConvertDQ.
 /// Matches FmhaBwdConvertQGradKernel::GridSize():
-///   dim3(ceil(seqlen_q / kM0), nhead, batch).
+///   GridDim(ceil(seqlen_q / kM0), nhead, batch).
 /// kM0 = 64 (tile rows along seqlen_q for 1D kernels), NOT block_size.
 /// Precondition: tile_m0 > 0, seqlen_q >= 0, batch > 0, nhead > 0.
-constexpr dim3 convert_dq_grid_size(int batch, int nhead, int seqlen_q, int tile_m0 = 64)
+constexpr GridDim convert_dq_grid_size(int batch, int nhead, int seqlen_q, int tile_m0 = 64)
 {
-    return dim3(static_cast<unsigned>((seqlen_q + tile_m0 - 1) / tile_m0),
-                static_cast<unsigned>(nhead),
-                static_cast<unsigned>(batch));
+    return {static_cast<unsigned>((seqlen_q + tile_m0 - 1) / tile_m0),
+            static_cast<unsigned>(nhead),
+            static_cast<unsigned>(batch)};
 }
 
 } // namespace rocm_ck
